This is a control script for analysis of PE reads
This is a control script that takes in PE Illumina reads and generates
most of the required analyses.  The script currently requires the PE
reads to be uniformly named as name.R1.fq.gz and name.R2.fq.gz.  For the
name, I suggest using the leading names of "diploid_" or "AD1_", "AD2_",
etc.

The script generates log files for each accession analyzed so that you
can see the general progress, determine completion of each step, and
give you a better indication of where to restart.

Ultimately, I plan to make this script modular, such that you can call
parts of it individually.  For the most part, provided you follow the
naming scheme that I have set forth in the script or alter it to your
own, you can comment out things that are not applicable or cannibalize
parts as you see fit.